dockerfusefs/requirements.txt
google-cloud
google-cloud-storage
fio
fusepy